requirements.txt
Django==3.0.3
line-bot-sdk==1.16.0
gunicorn==20.0.4
APScheduler==3.6.3
urllib3==1.25.8
dj-database-url==0.5.0
dj-static==0.0.6
psycopg2==2.8.5
pytz==2019.3
django-heroku==0.3.1
whitenoise==5.0.1